requirements.txt
opencv-python
matplotlib
tensorboardX
torchvision
pillow
pyyaml